Shapefile.UpdateAttributes:   Avoid deleting the attributes table if SaveAs is saving to the original location.

+++ DotSpatial.Data/Shapefile.cs
@@ -637,9 +637,15 @@ namespace DotSpatial.Data
             {
                 if (File.Exists(Attributes.Filename))
                 {
-                    if (File.Exists(newFile)) File.Delete(newFile);
-                    File.Copy(Attributes.Filename, newFile);
-                    Attributes.Filename = newFile;
+                    if (newFile.Equals(Attributes.Filename))
+                    {   // Already using existing file
+                    }
+                    else
+                    {
+                        if (File.Exists(newFile)) File.Delete(newFile);
+                        File.Copy(Attributes.Filename, newFile);
+                        Attributes.Filename = newFile;
+                    }
                     return;
                 }
             }
